+++ CHANGELOG.md
@@ -2,10 +2,20 @@
 
 ## Unreleased
 
+- Use the Node.js `require.cache` for caching the exports returned from a
+  Hook's `onrequire`. This allows users to delete entries from `require.cache`
+  to trigger a re-load (and re-run of the hook's `onrequire`) of a module the
+  next time it is required -- as mentioned at
+  https://nodejs.org/docs/latest/api/all.html#all_modules_requirecache
+  (https://github.com/elastic/require-in-the-middle/issues/61)
+
+- (SEMVER-MAJOR) Remove the `hook.cache` field. In earlier versions this was
+  available and some tests used it. However it was never a documented field.
+
 - If resolving the filename for a `require(...)` fails, defer to the wrapped
   require implementation rather than failing right away. This allows a
   possibly-monkey-patched `require` to do its own special thing.
-  https://github.com/elastic/require-in-the-middle/pull/59
+  (https://github.com/elastic/require-in-the-middle/pull/59)
 
 ## v6.0.0
 
@@ -21,12 +31,12 @@
 
 - Add support for hooking into the require of Node core modules prefixed with
   'node:', e.g. `require('node:http')`. See https://nodejs.org/api/modules.html#core-modules
-  https://github.com/elastic/require-in-the-middle/pull/53
+  (https://github.com/elastic/require-in-the-middle/pull/53)
 
 ## v5.1.0
 
 - Add support for hooking into require of absolute paths.
-  https://github.com/elastic/require-in-the-middle/issues/43
+  (https://github.com/elastic/require-in-the-middle/issues/43)
 
 ## earlier versions